Factor out redundant definitions of StatusTypeRelatedField
