+++ api/v2/serializers/details/machine_request.py
@@ -7,7 +7,6 @@ from core.models import (
     AtmosphereUser as User,
     IdentityMembership
 )
-from core.models.status_type import StatusType
 
 from api.v2.serializers.summaries import (
     AllocationSummarySerializer,
@@ -21,11 +20,10 @@ from api.v2.serializers.summaries import (
     ProviderSummarySerializer,
     ProviderMachineSummarySerializer,
     QuotaSummarySerializer,
-    UserSummarySerializer,
-    StatusTypeSummarySerializer
+    UserSummarySerializer
 )
 from api.v2.serializers.fields import (
-    ProviderMachineRelatedField, ModelRelatedField, IdentityRelatedField
+    ProviderMachineRelatedField, ModelRelatedField, IdentityRelatedField, StatusTypeRelatedField
 )
 from api.v2.serializers.fields.base import UUIDHyperlinkedIdentityField
 from api.validators import NoSpecialCharacters
@@ -55,39 +53,11 @@ class ProviderRelatedField(serializers.RelatedField):
 
     def get_queryset(self):
         return Provider.objects.all()
-    
+
     def to_representation(self, value):
         provider = Provider.objects.get(id=value.id)
         serializer = ProviderSummarySerializer(provider, context=self.context)
         return serializer.data
-
-
-class StatusTypeRelatedField(serializers.RelatedField):
-
-    def get_queryset(self):
-        return StatusType.objects.all()
-
-    def to_representation(self, value):
-        status_type = StatusType.objects.get(pk=value.pk)
-        serializer = StatusTypeSummarySerializer(
-            status_type,
-            context=self.context)
-        return serializer.data
-
-    def to_internal_value(self, data):
-        queryset = self.get_queryset()
-        if isinstance(data, dict):
-            identity = data.get("id", None)
-        else:
-            identity = data
-
-        try:
-            return queryset.get(id=identity)
-        except:
-            raise exceptions.ValidationError(
-                "StatusType with id '%s' does not exist."
-                % identity
-            )
 
 
 class MachineRequestSerializer(serializers.HyperlinkedModelSerializer):
@@ -107,9 +77,7 @@ class MachineRequestSerializer(serializers.HyperlinkedModelSerializer):
         queryset=Instance.objects.all(),
         serializer_class=InstanceSummarySerializer,
         style={'base_template': 'input.html'})
-    status = StatusTypeRelatedField(queryset=StatusType.objects.none(),
-                                    allow_null=True,
-                                    required=False)
+    status = StatusTypeRelatedField(allow_null=True, required=False)
     old_status = serializers.CharField(required = False)
 
     new_application_visibility = serializers.CharField()
@@ -212,9 +180,7 @@ class UserMachineRequestSerializer(serializers.HyperlinkedModelSerializer):
         style={'base_template': 'input.html'})
     start_date = serializers.DateTimeField(read_only=True)
     end_date = serializers.DateTimeField(read_only=True)
-    status = StatusTypeRelatedField(queryset=StatusType.objects.none(),
-                                    allow_null=True,
-                                    required=False)
+    status = StatusTypeRelatedField(allow_null=True, required=False)
     old_status = serializers.CharField(source='clean_old_status', required=False)
 
     new_application_visibility = serializers.CharField()
