+++ api/v2/serializers/fields/__init__.py
@@ -4,7 +4,8 @@ from .user import UserRelatedField
 from .tag import TagRelatedField
 from .base import ModelRelatedField
 from .identity import IdentityRelatedField
+from .status_type import StatusTypeRelatedField
 
 __all__ = ("ImageVersionRelatedField", "ProviderMachineRelatedField",
            "UserRelatedField", "ModelRelatedField", "TagRelatedField",
-           "IdentityRelatedField")
+           "IdentityRelatedField", "StatusTypeRelatedField")
